chore: Move reusable parts into shared. Keep only streaming parts in the summarize_session.
